Adding jQuery compatibility using componentDidMount method
