+++ .meteor/.finished-upgraders
@@ -13,3 +13,4 @@ notices-for-facebook-graph-api-2
 1.3.0-split-minifiers-package
 1.4.0-remove-old-dev-bundle-link
 1.4.1-add-shell-server-package
+1.4.3-split-account-service-packages
