+++ .meteor/packages
@@ -6,17 +6,17 @@
 
 meteor-base@1.0.4             # Packages every Meteor app needs to have
 mobile-experience@1.0.4       # Packages for a great mobile UX
-mongo@1.1.14                   # The database Meteor supports right now
+mongo@1.1.15                   # The database Meteor supports right now
 blaze-html-templates@1.0.4 # Compile .html files into Meteor Blaze views
 reactive-var@1.0.11            # Reactive variable for tracker
 jquery@1.11.10                  # Helpful client-side library
-tracker@1.1.1                 # Meteor's client-side reactive programming library
+tracker@1.1.2                 # Meteor's client-side reactive programming library
 
-standard-minifier-css@1.3.2   # CSS minifier run for production mode
-standard-minifier-js@1.2.1    # JS minifier run for production mode
+standard-minifier-css@1.3.3   # CSS minifier run for production mode
+standard-minifier-js@1.2.2    # JS minifier run for production mode
 es5-shim@4.6.15                # ECMAScript 5 compatibility for older browsers.
-ecmascript@0.6.1              # Enable ECMAScript2015+ syntax in app code
-shell-server@0.2.1            # Server-side component of the `meteor shell` command
+ecmascript@0.6.3              # Enable ECMAScript2015+ syntax in app code
+shell-server@0.2.2            # Server-side component of the `meteor shell` command
 
 fourseven:scss
 velocityjs:velocityjs
@@ -24,3 +24,4 @@ dispatch:mocha-phantomjs
 dispatch:mocha
 dispatch:phantomjs-tests@=0.0.5
 mdg:validated-method
+webtempest:animate
