+++ .meteor/versions
@@ -1,7 +1,7 @@
-aldeed:browser-tests@0.0.1
+aldeed:browser-tests@0.1.0
 allow-deny@1.0.5
-autoupdate@1.2.11
-babel-compiler@6.13.0
+autoupdate@1.3.12
+babel-compiler@6.14.1
 babel-runtime@1.0.1
 base64@1.0.10
 binary-heap@1.0.10
@@ -12,17 +12,18 @@ boilerplate-generator@1.0.11
 caching-compiler@1.1.9
 caching-html-compiler@1.1.0
 callback-hook@1.0.10
-check@1.2.4
+check@1.2.5
+coffeescript@1.0.17
 ddp@1.2.5
-ddp-client@1.2.9
+ddp-client@1.3.3
 ddp-common@1.2.8
-ddp-server@1.2.10
+ddp-server@1.3.13
 deps@1.0.12
 diff-sequence@1.0.7
-dispatch:mocha@0.1.0
+dispatch:mocha@0.2.0
 dispatch:mocha-phantomjs@0.1.7
 dispatch:phantomjs-tests@0.0.5
-ecmascript@0.6.1
+ecmascript@0.6.3
 ecmascript-runtime@0.3.15
 ejson@1.0.13
 es5-shim@4.6.15
@@ -32,26 +33,26 @@ geojson-utils@1.0.10
 hot-code-push@1.0.4
 html-tools@1.0.11
 htmljs@1.0.11
-http@1.1.8
+http@1.2.11
 id-map@1.0.9
 jquery@1.11.10
-launch-screen@1.0.12
+launch-screen@1.1.1
 livedata@1.0.18
-logging@1.1.16
+logging@1.1.17
 mdg:validated-method@1.1.0
-meteor@1.6.0
+meteor@1.6.1
 meteor-base@1.0.4
-minifier-css@1.2.15
-minifier-js@1.2.15
-minimongo@1.0.19
+minifier-css@1.2.16
+minifier-js@1.2.17
+minimongo@1.0.20
 mobile-experience@1.0.4
-mobile-status-bar@1.0.13
-modules@0.7.7
-modules-runtime@0.7.8
-mongo@1.1.14
+mobile-status-bar@1.0.14
+modules@0.7.9
+modules-runtime@0.7.9
+mongo@1.1.15
 mongo-id@1.0.6
 npm-mongo@2.2.16_1
-observe-sequence@1.0.14
+observe-sequence@1.0.15
 ordered-dict@1.0.9
 practicalmeteor:mocha-core@1.0.1
 promise@0.8.8
@@ -60,19 +61,20 @@ reactive-var@1.0.11
 reload@1.1.11
 retry@1.0.9
 routepolicy@1.0.12
-shell-server@0.2.1
+shell-server@0.2.2
 spacebars@1.0.13
 spacebars-compiler@1.1.0
-standard-minifier-css@1.3.2
-standard-minifier-js@1.2.1
+standard-minifier-css@1.3.3
+standard-minifier-js@1.2.2
 templating@1.3.0
 templating-compiler@1.3.0
 templating-runtime@1.3.0
 templating-tools@1.1.0
-tracker@1.1.1
+tracker@1.1.2
 ui@1.0.12
 underscore@1.0.10
-url@1.0.11
+url@1.1.0
 velocityjs:velocityjs@1.2.1
-webapp@1.3.12
+webapp@1.3.13
 webapp-hashing@1.0.9
+webtempest:animate@0.1.9
