+++ package.json
@@ -60,6 +60,7 @@
   "dependencies": {
     "arrayof": "^0.1.0",
     "babel-runtime": "^6.18.0",
+    "bcrypt": "^1.0.2",
     "body-parser": "^1.15.2",
     "chai": "^3.5.0",
     "express": "^4.14.0",
@@ -89,6 +90,7 @@
     "eslint-plugin-jsx-a11y": "^2.2.2",
     "eslint-plugin-meteor": "^4.0.0",
     "eslint-plugin-react": "^6.2.1",
+    "mocha": "^3.2.0",
     "node-sass": "^4.5.0",
     "style-loader": "^0.13.1"
   }
